+++ cmd/otel-allocator/internal/telemetry/delta_producer_test.go
@@ -1,7 +1,7 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
-package main
+package telemetry
 
 import (
 	"context"
